+++ backend/.gitignore
@@ -7,4 +7,5 @@ pyvenv.cfg
 **/__pycache__/
 **/__init__.py
 .env
-.vscode/
+.vscode/bin/
+
