--- backend/announcements/views.py
@@ -1,11 +1,0 @@
-from rest_framework import generics
-from .models import Announcement
-from .serializers import AnnouncementSerializer
-
-class AnnouncementListCreate(generics.ListCreateAPIView):
-    queryset = Announcement.objects.all()
-    serializer_class = AnnouncementSerializer
-
-class AnnouncementDetail(generics.RetrieveUpdateDestroyAPIView):
-    queryset = Announcement.objects.all()
-    serializer_class = AnnouncementSerializer
